add option for delete of crd to utility function

+++ platform-operator/scripts/uninstall/uninstall-utils.sh
@@ -46,6 +46,7 @@ function err_return() {
 
 # Deletes kubernetes resources from all namespaces
 # $1 resource-type - type of the resources being deleted
+# $2 crd-delete-flag - value of "yes" or "no" to also delete crd, default is "yes" for compatibility
 function delete_k8s_resource_from_all_namespaces() {
   local res=$1
   if kubectl get crd "${res}"> /dev/null 2>&1 ; then
@@ -57,8 +58,10 @@ function delete_k8s_resource_from_all_namespaces() {
     done
   fi
   # Delete the CRDs, without any CRs based on that, from all the namespaces
-  if kubectl get crd "${res}"> /dev/null 2>&1 ; then
-    kubectl delete crd "${res}" --ignore-not-found > /dev/null 2>&1
+  if [ -z "$2" ] || [ "$2" == "yes" ]; then
+    if kubectl get crd "${res}"> /dev/null 2>&1 ; then
+      kubectl delete crd "${res}" --ignore-not-found > /dev/null 2>&1
+    fi
   fi
 }
